pyproject.toml
[project]
name = "dxtr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.0.0",
  "httpx>=0.25.0",
  "fastapi>=0.128.0",
  "uvicorn>=0.30.0",
  "pydantic-ai==1.12.0",
  "pydantic-ai-litellm>=0.2.3",
  "psycopg2-binary>=2.9.11",
  "python-dotenv>=1.2.1",
  "google-cloud-storage>=3.8.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dxtr"] # Or ["src/dxtr"] if using the src layout
